Use aria-describedby to announce current value
Use aria-describedby instead of aria-activedescendant to announce the
current value of the field. For some unknown reason,
aria-activedescendant does not work with NVDA and Firefox for single
valued selects (perhaps because the <input> and value display overlap?).

The downside is that aria-describedby is not currently supported by
ChromeVox or Android Talkback (see
http://davidmacd.com/blog/is-tabindex-required-on-describedby.html
).