Stop screen readers from reading close boxes
Previously, close boxes were rendered by using "&times;" with an
aria-label of "clear value" or similar. This was causing screen
readers to say "clear value" or "times" while reading the
contents of a combobox.

Now, we will change aria-label in the relevant places to be
a comma. This will cause screen readers not to say "clear value"
or "times" when reading the contents of a combobox when the
close box is visible.

It should be noted that this option is something of a last resort.
The following things were tried, and still resulted in the screen
reader saying "clear value" or "times" when reading the combobox's
contents:
-Using aria-hidden="true" on the element including &times, or on
  its parent element, or both
-Using aria-label="" and/or title="" on the span including &times,
  or on its parent element, or both
-Using aria-labelledby set to a separate element that was blank
  on the span including &times, or on its parent element, or both

Use of aria-label="," will result in a slight vocal pause, but
this seems less bad than saying "clear value" or "times" for
a reason that will not be discernible to the user of a screen
reader. Even the worst-case scenario that the screen reader
says the word "comma" (which I didn't observe, but probably
could happen) doesn't seem worse than the current behavior.

Story: R-15476
Reviewed-by: pbrant